+++ packages/core/src/exports/internal.ts
@@ -1,4 +1,5 @@
 import { Directory } from '../core/dir.js';
 import { LoomFile } from '../core/file.js';
+import { LineResult } from '../helper/result.js';
 
-export { Directory, LoomFile };
+export { Directory, LoomFile, LineResult };
