+++ plugins/front-matter-converter/src/core/parse.ts
@@ -1,0 +1,51 @@
+import { LoomFile } from '@loom-io/core';
+import { LineResult } from '@loom-io/core/internal';
+import { getFrontMatterConverter, hasFrontMatter } from './utils';
+
+
+export async function readFrontMatter(line: LineResult) {
+	const frontMatter = [];
+	let lineContent = await line.read('utf8');
+	while(lineContent !== '---') {
+		frontMatter.push(lineContent);
+		await line.next();
+		lineContent = await line.read('utf-8');
+	}
+	return frontMatter.join('\n');
+}
+
+export async function readContent(line: LineResult) {
+	const content = [];
+	do {
+		const lineContent = await line.read('utf8');
+		if(!(content.length === 0 && (lineContent === '---' || lineContent === ''))) {
+			content.push(lineContent);
+		}
+		await line.next();
+	} while(await line.hasNext());
+	return content.join('\n');
+}
+
+export async function parse(file: LoomFile): Promise<unknown> {
+	const reader = await file.reader();
+	const lineReader = await reader.firstLine();
+
+	const result = {
+		data: {} as unknown,
+		content: ''
+	};
+
+
+	if(await hasFrontMatter(lineReader)) {
+		lineReader.next();
+		const frontMatter = await readFrontMatter(lineReader);
+		const { parse } = await getFrontMatterConverter(lineReader);
+		result.data = await parse(frontMatter);
+
+		await lineReader.next();
+	}
+
+	result.content = await readContent(lineReader);
+
+	return result;
+}
